+++ pkg/azuremachine/validate_update.go
@@ -12,6 +12,7 @@ import (
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
 	"github.com/giantswarm/azure-admission-controller/internal/releaseversion"
 	"github.com/giantswarm/azure-admission-controller/internal/semverhelper"
+	"github.com/giantswarm/azure-admission-controller/pkg/generic"
 	"github.com/giantswarm/azure-admission-controller/pkg/validator"
 )
 
@@ -52,6 +53,11 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 	}
 
 	err := checkSSHKeyIsEmpty(ctx, azureMachineNewCR)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
+	err = generic.ValidateOrganizationLabelUnchanged(azureMachineOldCR, azureMachineNewCR)
 	if err != nil {
 		return false, microerror.Mask(err)
 	}