+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Prevent Organization label value change on CR updates.
+
 ### Removed
 
 - Removed Organization label value normalization on CR creation.